+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.0] - 2020-06-17
+
 ### Added
 
 - `Logs`: supports time and usage metrics
@@ -270,7 +272,8 @@ First stable release.
 Please see the [Releases](https://github.com/Open-EO/openeo-vue-components/releases) for changelogs prior to v1.0.0.
 
 
-[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.1.2...HEAD
+[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.2.0...HEAD
+[2.2.0]: https://github.com/Open-EO/openeo-vue-components/compare/v2.1.2...v2.2.0
 [2.1.2]: https://github.com/Open-EO/openeo-vue-components/compare/v2.1.1...v2.1.2
 [2.1.1]: https://github.com/Open-EO/openeo-vue-components/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/Open-EO/openeo-vue-components/compare/v2.0.3...v2.1.0